Move articles list to the separate component.
