+++ src/main.ts
@@ -4,6 +4,7 @@ import router from './router';
 import Vuetify from 'vuetify';
 import Paginate from 'vuejs-paginate';
 import VueMarkdown from 'vue-markdown';
+import ArticlesList from "@/components/ArticlesList.vue";
 import 'vuetify/dist/vuetify.min.css';
 import 'material-design-icons-iconfont/dist/material-design-icons.css';
 
@@ -20,3 +21,4 @@ Vue.use(Vuetify, {
 
 Vue.component('paginate', Paginate);
 Vue.component('vue-markdown', VueMarkdown);
+Vue.component('articles-list', ArticlesList);
