+++ src/views/Articles.vue
@@ -1,36 +1,11 @@
 <template>
-  <div class="articles">
-    <div v-for="article in paginatedArticles" :key="article.title" style="border: 1px solid gray">
-      <p>Tytuł: {{ article.title }}</p>
-      <p>Autor: {{article.creator.user.username}}, Data: {{article.creation_date}}</p>
-      <p>
-        Tagi:
-        <span
-          v-for="articleTag in article.tags"
-          :key="articleTag.tag.name"
-        >#{{ articleTag.tag.name }}</span>
-      </p>
-      <p>Treść:
-        <vue-markdown>{{article.text}}</vue-markdown>
-      </p>
-    </div>
-
-    <paginate
-      v-if="pagination != undefined"
-      :page-count="pagination.pageCount"
-      :page-range="3"
-      :margin-pages="2"
-      :click-handler="paginationClicked"
-      :prev-text="'Poprzednia strona'"
-      :next-text="'Następna strona'"
-      :container-class="'paginationContainer'"
-    ></paginate>
-  </div>
+  <articles-list v-bind:articles="articles" v-bind:pagination="pagination"></articles-list>
 </template>
 
 <script lang="ts">
 import { Component, Vue, Prop } from "vue-property-decorator";
 import HelloWorld from "@/components/HelloWorld.vue";
+import ArticlesList from "@/components/ArticlesList.vue";
 import { ArticlesService } from "@/services/ArticlesService";
 import { ArticleModel } from "@/models/ArticleModel";
 import { PaginationModel } from "@/models/PaginationModel";
@@ -42,11 +17,8 @@ import { PaginationModel } from "@/models/PaginationModel";
 })
 export default class Articles extends Vue {
   private articlesService!: ArticlesService;
-  private allArticles!: ArticleModel[];
-  private paginatedArticles!: ArticleModel[];
+  private articles!: ArticleModel[];
   private pagination!: PaginationModel;
-
-  private articlesPerPage: number = 4;
 
   constructor() {
     super();
@@ -55,28 +27,16 @@ export default class Articles extends Vue {
   beforeCreate() {
     this.articlesService = new ArticlesService();
     this.articlesService.getAllArticles().then(articles => {
-      this.allArticles = articles;
-      this.pagination = new PaginationModel(
-        articles.length,
-        this.articlesPerPage
-      );
-
-      this.paginationClicked(1);
+      this.articles = articles;
+      this.pagination = new PaginationModel(articles.length, 4);
     });
   }
 
   public data() {
     return {
       pagination: this.pagination,
-      paginatedArticles: this.paginatedArticles
+      articles: this.articles
     };
-  }
-
-  public paginationClicked(pageNumber: number) {
-    this.paginatedArticles = this.allArticles.slice(
-      (pageNumber - 1) * this.articlesPerPage,
-      pageNumber * this.articlesPerPage
-    );
   }
 }
 </script>
